Upgrade Example (Android) App To 0.60.x
Summary:
I tried runing the sample app but it was no longer
compiling due to the library having been upgraded
to 0.60.x and thus Androidx breaking it.

This updates the Android part of the example.

It adds jetify as a post install task to fix up the
gesture handler lib.